move encoding tests to src/stream.rs
