+++ helix-view/src/stream.rs
@@ -160,3 +160,73 @@ where
     writer.flush().await?;
     Ok(())
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use std::path::PathBuf;
+
+    macro_rules! decode {
+        ($name:ident, $label:expr, $label_override:expr) => {
+            #[test]
+            fn $name() {
+                let encoding = encoding::Encoding::for_label($label_override.as_bytes()).unwrap();
+                let base_path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("tests/encoding");
+                let path = base_path.join(format!("{}_in.txt", $label));
+                let ref_path = base_path.join(format!("{}_in_ref.txt", $label));
+                assert!(path.exists());
+                assert!(ref_path.exists());
+
+                let mut file = std::fs::File::open(path).unwrap();
+                let text = from_reader(&mut file, RopeWrite::default(), Some(encoding))
+                    .unwrap()
+                    .0
+                    .finish()
+                    .to_string();
+                let expectation = std::fs::read_to_string(ref_path).unwrap();
+                assert_eq!(text[..], expectation[..]);
+            }
+        };
+        ($name:ident, $label:expr) => {
+            decode!($name, $label, $label);
+        };
+    }
+
+    macro_rules! encode {
+        ($name:ident, $label:expr, $label_override:expr) => {
+            #[test]
+            fn $name() {
+                let encoding = encoding::Encoding::for_label($label_override.as_bytes()).unwrap();
+                let base_path = PathBuf::from(env!("CARGO_MANIFEST_DIR")).join("tests/encoding");
+                let path = base_path.join(format!("{}_out.txt", $label));
+                let ref_path = base_path.join(format!("{}_out_ref.txt", $label));
+                assert!(path.exists());
+                assert!(ref_path.exists());
+
+                let text = Rope::from_str(&std::fs::read_to_string(path).unwrap());
+                let mut buf: Vec<u8> = Vec::new();
+                helix_lsp::block_on(to_writer(&mut buf, encoding, text.chunks())).unwrap();
+
+                let expectation = std::fs::read(ref_path).unwrap();
+                assert_eq!(buf, expectation);
+            }
+        };
+        ($name:ident, $label:expr) => {
+            encode!($name, $label, $label);
+        };
+    }
+
+    decode!(big5_decode, "big5");
+    encode!(big5_encode, "big5");
+    decode!(euc_kr_decode, "euc_kr", "EUC-KR");
+    encode!(euc_kr_encode, "euc_kr", "EUC-KR");
+    decode!(gb18030_decode, "gb18030");
+    encode!(gb18030_encode, "gb18030");
+    decode!(iso_2022_jp_decode, "iso_2022_jp", "ISO-2022-JP");
+    encode!(iso_2022_jp_encode, "iso_2022_jp", "ISO-2022-JP");
+    decode!(jis0208_decode, "jis0208", "EUC-JP");
+    encode!(jis0208_encode, "jis0208", "EUC-JP");
+    decode!(jis0212_decode, "jis0212", "EUC-JP");
+    decode!(shift_jis_decode, "shift_jis");
+    encode!(shift_jis_encode, "shift_jis");
+}
